Separated Camera class, Added Delegate pattern to access ViewController variables from custom classes
The Camera class was too large, so it has been split into the
CameraCapture and CameraPreview. The ViewController will no longer be
passed in as an init parameter on custom classes that require access to
it. I am now using the Delegate Pattern so that custom classes can have
access to specific portions of the ViewController. UML diagram updated.