SpeechSynthesizer complete. Text Finder complete. Updated ReadTextMachine
The SpeechSynthesizer is complete. I want to add a better voice
selection to it later, but it’s working properly (1st real world use
case of overloading in Swift for me).

The TextFinder is using Apple’s Vision Framework to detect text and
posting a notification when done. BoundingBoxes are saved inside the
object for use in the next states.