New state machine
Reviewed the state machine and drew a ReadTextMachine diagram to make
it easier to follow. Still buggy in the mid of cancellation and I
believe it’s because of the fact that I cannot cancel a triggered
Vision Request (on TextFinder). I’ll add the state checks after
notifications to see if we get something better.