Added state check prior to executing a new state.
Disabled screen tap during Vision Request processing. As it processes
rather fast, it shouldn’t be a problem. The reason I did it was because
I couldn’t find a request cancelation routine.

Facing the same issue with Tesseract. Even though I call out the cancel
request routine, it sends me the notification stating it has completed
a request afterwards. To solve it for now, I have placed a state check
prior to every internal state call. This reduces the probability of
incorrect execution, but I still need to figure out how to fully cancel
requests. Setting objects to nil doesn’t work.

The overall flow of the app is in much better shape, but I occasionally
get  to a black static screen when no text is detected. I can recover
by pressing the home button and reopening the app, so the AppDelegate
is doing its job.